@@ -3,15 +3,15 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "2.0.5"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "CI=true react-scripts test --env=jest-environment-jsdom-sixteen",
-    "watch": "react-scripts test --env=jest-environment-jsdom-sixteen --verbose=false",
+    "test": "CI=true react-scripts test",
+    "watch": "react-scripts test --verbose=false",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -24,8 +24,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.2",
-    "@testing-library/react": "^10.4.7",
-    "jest-environment-jsdom-sixteen": "^1.0.3"
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1"
   }
 }
