@@ -1,14 +1,8 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Icon from './Icon';
 import './Divider.css';
 
-class Divider extends Component {
-    render() {
-        const { name, code } = this.props;
-        return (
-            <h3 role="separator" className="card-divider">{name} <Icon code={code}/></h3>
-        )
-    }
-}
+const Divider = ({ name, code }) =>
+    <h3 role="separator" className="card-divider">{name} <Icon code={code}/></h3>;
 
 export default Divider;
