@@ -1,9 +1,7 @@
-import React, { Component } from 'react';
+import React from 'react';
 
-class FilterNotifier extends Component {
-    render() {
-        return this.props.on ? <span role="alert">&bull;</span> : ""
-    }
+const FilterNotifier = ({ on }) =>  {
+    return on ? <span role="alert">&bull;</span> : ""
 }
 
 export default FilterNotifier;
