@@ -1,15 +1,10 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './Header.css';
 
-class Header extends Component {
-    render() {
-        return (
-            <div id="title">
-                <h1>BlacKat</h1>
-                <span id="subtitle">Netrunner Card Explorer</span>
-            </div>
-        );
-    }
-}
+const Header = () => 
+    <div id="title">
+        <h1>BlacKat</h1>
+        <span id="subtitle">Netrunner Card Explorer</span>
+    </div>;
 
 export default Header;
