@@ -1,16 +1,11 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './Loader.css';
 
-class Loader extends Component {
-    render() {
-        return (
-            <div className="spinner" role="progressbar">
-                <div className="bounce1"></div>
-                <div className="bounce2"></div>
-                <div className="bounce3"></div>
-            </div>
-        )
-    }
-}
+const Loader = () => 
+    <div className="spinner" role="progressbar">
+        <div className="bounce1"></div>
+        <div className="bounce2"></div>
+        <div className="bounce3"></div>
+    </div>;
 
 export default Loader;
