@@ -1,25 +1,22 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './TextSearch.css';
 
-class TextSearch extends Component {
-    static defaultProps = {
-        placeholder: "search",
-        value: ""
-    }
+const TextSearch = ({
+        placeholder = "search",
+        value = "",
+        onChange
+    }) => {
 
-    handleChange = ev => {
+    const handleChange = ev => {
         ev.preventDefault();
-        this.props.onChange(ev.target.value);
+        onChange(ev.target.value);
     }
 
-    render() {
-        const { value, placeholder } = this.props;
-        return (
-            <div className="form-group">
-                <input className="form-control" defaultValue={value} placeholder={placeholder} onInput={this.handleChange} />
-            </div>
-        );
-    }
+    return (
+        <div className="form-group">
+            <input className="form-control" defaultValue={value} placeholder={placeholder} onInput={handleChange} />
+        </div>
+    );
 }
 
 
